Update package.json
Today, this is pulling in an older version of NodeJS14 which is no longer supported.  So, when you go to deploy solution CDK will error out when trying to spin up auto generated lambda's with DynamoDB cross-region support.  By changing this version you will now be generating lambda's using NodeJS20